+++ lib/change_charity_components.dart
@@ -3,3 +3,4 @@ library change_charity_components;
 export 'paintings.dart';
 export 'submit_row.dart';
 export 'text_field.dart';
+export 'theme.dart';