+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## [0.0.3] - Paintings, SubmitRow, and TextField all imported through change_charity_components.dart
+## [0.0.8] - small fixes
+
+## [0.0.7] - added themes, and dark mode
+
+## [0.0.6] - added autofill hints
+
+## [0.0.5] - added null assertions, ability to change submit button color
 
 ## [0.0.4] - added example
+
+## [0.0.3] - Paintings, SubmitRow, and TextField all imported through change_charity_components.dart